requirements.txt
ftfy==6.1.1
matplotlib==3.5.2
omegaconf==2.2.0
pandas==1.4.2
pytorch-lightning==1.6.3
PyYAML==6.0
regex==2022.4.24
toml==0.10.2
numpy
torch
torchvision
ipython
tqdm==4.64.0
openpyxl==3.0.10
